These are the files.
These have images to help you people not search the dark web for a few images and probably get a trojan(Eg. Bearfoos) or even a Mydoom. Please use those images as I can give you 100% guarantee they are safe and do not be afraid of copyright of programs as they are created by me.